Separate tests to SqsUnboundedReaderTest and SqsUnboundedSourceTest
